Add "tableExists" method to MigrationAssistant.

+++ brightify-core/src/main/java/com/brightgestures/brightify/util/MigrationAssistantImpl.java
@@ -1,5 +1,6 @@
 package com.brightgestures.brightify.util;
 
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import com.brightgestures.brightify.BrightifyFactory;
 import com.brightgestures.brightify.EntityMetadata;
@@ -20,21 +21,39 @@ public class MigrationAssistantImpl<ENTITY> implements MigrationAssistant<ENTITY
 
     @Override
     public void createTable() {
-        entityMetadata.createTable(getDatabase());
+        if(!tableExists()) {
+            entityMetadata.createTable(getDatabase());
+        }
     }
 
     @Override
     public void dropTable() {
-        DropTable dropTable = new DropTable();
-        dropTable.setTableName(entityMetadata.getTableName());
-        dropTable.setDatabaseName(brightifyFactory.getDatabaseName());
-        dropTable.run(getDatabase());
+        if(tableExists()) {
+            DropTable dropTable = new DropTable();
+            dropTable.setTableName(entityMetadata.getTableName());
+            dropTable.setDatabaseName(brightifyFactory.getDatabaseName());
+            dropTable.run(getDatabase());
+        }
     }
 
     @Override
     public void dropCreateTable() {
         dropTable();
         createTable();
+    }
+
+    @Override
+    public boolean tableExists() {
+        Cursor cursor = getDatabase().rawQuery("SELECT DISTINCT tbl_name FROM sqlite_master WHERE tbl_name = ?",
+                                      new String[] { entityMetadata.getTableName() });
+        if(cursor != null) {
+            if(cursor.getCount() > 0) {
+                cursor.close();
+                return true;
+            }
+            cursor.close();
+        }
+        return false;
     }
 
     private SQLiteDatabase getDatabase() {
